requirements.txt
tavily-python
python-dotenv
streamlit